% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dge_list}
\alias{dge_list}
\title{dge_list}
\format{
A DGEList object including the GSE63310 dataset
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63310}
}
\usage{
data(dge_list)
}
\description{
dge_list  gene count data in SummarizedExperiment format
}
\keyword{datasets}
