% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_experiment}
\alias{single_cell_experiment}
\title{single_cell_experiment}
\format{
A SingleCellExperiment object
}
\source{
\url{https://osca.bioconductor.org}
}
\usage{
data(single_cell_experiment)
}
\description{
dge_list  gene count data in SingleCellExperiment format
}
\keyword{datasets}
